fix: move contract types to `/src`
